devops_app/requirements.txt
flask
gunicorn
pycryptodome
Authlib
functools